Separated different phases of logic into different files
