+++ OpenGLApp/OpenGLApp/Main.hpp
@@ -1,0 +1,26 @@
+#ifndef _MAIN_H
+#define _MAIN_H
+
+// clang-format off
+#include <glad/glad.h>
+#include <GLFW/glfw3.h>
+// clang-format on
+
+#include <iostream>
+#include <list>
+
+#include "./utils.hpp"
+#include "./shader_s.h"
+#include "asteroids/spaceship.hpp"
+#include "asteroids/projectile.hpp"
+#include "asteroids/asteroid.hpp"
+#include "asteroids/star.hpp"
+#include "asteroids/displayString.hpp"
+#include "asteroids/_debugOpts.hpp"
+#include "asteroids/heart.hpp"
+
+#include "asteroids/game_logic/game.hpp"
+
+enum gamePhases { mainMenu, instructions, game, endScreen };
+
+#endif
